Simple position column with skeleton
