+++ src/frontend/src/pages/ProjectPage/components/PositionColumn/styles.less
@@ -1,0 +1,19 @@
+.container {
+    display: flex;
+    align-items: center;
+    flex-grow: 1;
+
+    .details {
+        margin-left: calc(var(--grid-unit) * 2px);
+        flex-grow: 1;
+    }
+
+    .positionName {
+        font-weight: bold;
+        margin-bottom: calc(var(--grid-unit) * .5px);
+    }
+
+    .personName {
+        font-size: 12px;
+    }
+}
